feat: use a web worker to generate client packages
