+++ src/utils/pack-worker-client.ts
@@ -1,0 +1,3 @@
+import PackerWorker from './pack-worker?worker'
+
+export const packerWorker = new PackerWorker()
